+++ frontend/packages/forms/src/index.ts
@@ -2,6 +2,7 @@ export * from "./ChangePasswordForm.js";
 export * from "./ForgotPasswordForm.js";
 export * from "./i18n.js";
 export * from "./LoginForm.js";
+export * from "./LegalConsent.js";
 export * from "./OAuthButtons.js";
 export * from "./RegisterForm.js";
 export * from "./ResetPasswordForm.js";